% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in_sql.R
\name{csv_in_sql}
\alias{csv_in_sql}
\title{Embed CSV data in SQL}
\usage{
csv_in_sql(data, tab_name = "tbl", separator = "|")
}
\arguments{
\item{data}{dataset to be translated}

\item{tab_name}{name of the resulting "WITH table".}

\item{separator}{specifies string to be used as delimiter in the SQL query (use any character/string that is not present in the data).}
}
\value{
SQL query as a character value.
}
\description{
Translate a dataset into a plaintext SQL representation
}
\details{
Tested with PL/SQL.
}
\examples{
# data(iris)
# cat(csv_in_sql(iris))
}
